backend/requirements.txt
fastapi
uvicorn
pandas
requests
beautifulsoup4
yfinance
python-dotenv
httpx